--- a/lib/model.js
+++ b/lib/model.js
@@ -1703,7 +1703,7 @@ Model.findByIdAndRemove = function(id, options, callback) {
  *       // ...
  *     })
  *
- * @param {Array|Object...} doc(s)
+ * @param {Array|Object|*} doc(s)
  * @param {Function} [callback] callback
  * @return {Promise}
  * @api public
@@ -2026,7 +2026,7 @@ Model.mapReduce = function mapReduce(o, callback) {
  *        console.log(results);
  *     });
  *
- * @param {Object/Array} GeoJSON point or legacy coordinate pair [x,y] to search near
+ * @param {Object|Array} GeoJSON point or legacy coordinate pair [x,y] to search near
  * @param {Object} options for the qurery
  * @param {Function} [callback] optional callback for the query
  * @return {Promise}