--- a/test/cast.test.js
+++ b/test/cast.test.js
@@ -8,29 +8,33 @@ var cast = require('../lib/cast');
 
 describe('cast: ', function() {
   describe('bitwise query operators: ', function() {
-    it('with a number', function() {
+    it('with a number', function(done) {
       var schema = new Schema({ x: Buffer });
       assert.deepEqual(cast(schema, { x: { $bitsAllClear: 3 } }),
         { x: { $bitsAllClear: 3 } });
+      done();
     });
 
-    it('with an array', function() {
+    it('with an array', function(done) {
       var schema = new Schema({ x: Buffer });
       assert.deepEqual(cast(schema, { x: { $bitsAllSet: [2, '3'] } }),
         { x: { $bitsAllSet: [2, 3] } });
+      done();
     });
 
-    it('with a buffer', function() {
+    it('with a buffer', function(done) {
       var schema = new Schema({ x: Number });
       assert.deepEqual(cast(schema, { x: { $bitsAnyClear: new Buffer([3]) } }),
         { x: { $bitsAnyClear: new Buffer([3]) } });
+      done();
     });
 
-    it('throws when invalid', function() {
+    it('throws when invalid', function(done) {
       var schema = new Schema({ x: Number });
       assert.throws(function() {
         cast(schema, { x: { $bitsAnySet: 'Not a number' } });
       }, /Cast to number failed/);
+      done();
     });
   });
 });