tests fixed for build with makefile which uses --async-only option
js docs fixed for new `dox` version