@TrejGun 2016-01-26
benchmark

before
```
running string
string took 7.114 seconds for 10000 docs (1405.678942929435 dps)
change:  { rss: 36605952, heapTotal: 35085824, heapUsed: 20202976 }
running number
number took 7.569 seconds for 10000 docs (1321.178491214163 dps)
change:  { rss: 352256, heapTotal: 0, heapUsed: -6827776 }
running date
date took 7.42 seconds for 10000 docs (1347.7088948787061 dps)
change:  { rss: 1048576, heapTotal: 1031936, heapUsed: 3687008 }
running bool
bool took 7.43 seconds for 10000 docs (1345.8950201884254 dps)
change:  { rss: 0, heapTotal: 0, heapUsed: -12870312 }
running buffer
buffer took 7.476 seconds for 10000 docs (1337.613697164259 dps)
change:  { rss: -532480, heapTotal: -1031936, heapUsed: 7777464 }
running objectid
objectid took 7.253 seconds for 10000 docs (1378.7398317937404 dps)
change:  { rss: 1056768, heapTotal: 1031936, heapUsed: -7871128 }
running array of mixed
array of mixed took 8.328 seconds for 10000 docs (1200.7684918347743 dps)
change:  { rss: 397312, heapTotal: 0, heapUsed: 3125720 }
running array of strings
array of strings took 8.15 seconds for 10000 docs (1226.993865030675 dps)
change:  { rss: 0, heapTotal: 0, heapUsed: 109128 }
running array of numbers
array of numbers took 8.37 seconds for 10000 docs (1194.7431302270013 dps)
change:  { rss: 81920, heapTotal: 0, heapUsed: 3304240 }
running array of dates
array of dates took 8.844 seconds for 10000 docs (1130.7100859339666 dps)
change:  { rss: 0, heapTotal: 0, heapUsed: -6327144 }
running array of bools
array of bools took 8.335 seconds for 10000 docs (1199.7600479904017 dps)
change:  { rss: 0, heapTotal: 0, heapUsed: -369584 }
running array of buffers
array of buffers took 8.596 seconds for 10000 docs (1163.3317822242905 dps)
change:  { rss: 4096, heapTotal: 0, heapUsed: -1525608 }
running array of objectids
array of objectids took 8.522 seconds for 10000 docs (1173.4334663224595 dps)
change:  { rss: 0, heapTotal: 0, heapUsed: 11034504 }
running array of docs
array of docs took 9.745 seconds for 10000 docs (1026.167265264238 dps)
change:  { rss: 376832, heapTotal: 1031936, heapUsed: 2184312 }
completed 140000 docs in 113.152 seconds (1237.2737556561085 dps)
```

after
```
running string
string took 6.974 seconds for 10000 docs (1433.8973329509606 dps)
change:  { rss: 36364288, heapTotal: 34065920, heapUsed: 19437240 }
running number
number took 6.91 seconds for 10000 docs (1447.178002894356 dps)
change:  { rss: 339968, heapTotal: 0, heapUsed: -7061304 }
running date
date took 7.366 seconds for 10000 docs (1357.588922074396 dps)
change:  { rss: 1048576, heapTotal: 1031936, heapUsed: 1207960 }
running bool
bool took 6.851 seconds for 10000 docs (1459.6409283316304 dps)
change:  { rss: 0, heapTotal: 0, heapUsed: -854376 }
running buffer
buffer took 7.19 seconds for 10000 docs (1390.8205841446452 dps)
change:  { rss: 536576, heapTotal: 0, heapUsed: 7393040 }
running objectid
objectid took 6.946 seconds for 10000 docs (1439.677512237259 dps)
change:  { rss: 16384, heapTotal: 0, heapUsed: -6588360 }
running array of mixed
array of mixed took 7.921 seconds for 10000 docs (1262.4668602449185 dps)
change:  { rss: 286720, heapTotal: 0, heapUsed: 2535456 }
running array of strings
array of strings took 7.798 seconds for 10000 docs (1282.3800974608873 dps)
change:  { rss: 0, heapTotal: 0, heapUsed: -99744 }
running array of numbers
array of numbers took 7.802 seconds for 10000 docs (1281.722635221738 dps)
change:  { rss: 28672, heapTotal: 0, heapUsed: -12349144 }
running array of dates
array of dates took 7.844 seconds for 10000 docs (1274.859765425803 dps)
change:  { rss: 8192, heapTotal: 0, heapUsed: 7429144 }
running array of bools
array of bools took 7.886 seconds for 10000 docs (1268.06999746386 dps)
change:  { rss: 81920, heapTotal: 0, heapUsed: -747944 }
running array of buffers
array of buffers took 8.221 seconds for 10000 docs (1216.3970319912419 dps)
change:  { rss: 77824, heapTotal: 0, heapUsed: -2570072 }
running array of objectids
array of objectids took 7.934 seconds for 10000 docs (1260.3982858583313 dps)
change:  { rss: 8192, heapTotal: 0, heapUsed: -5425512 }
running array of docs
array of docs took 8.648 seconds for 10000 docs (1156.3367252543942 dps)
change:  { rss: 1339392, heapTotal: 1031936, heapUsed: -1484928 }
completed 140000 docs in 106.29099999999998 seconds (1317.1387982049282 dps)
```