@TrejGun 2016-01-26
unfortunately i was unable to test `shards` and `replica sets`
can you please help me to set it up
i have no idea what to do with mongo