@vkarpov15 2016-01-26
Thanks. We usually don't use `make test` anymore though, prefer `npm test`. Why are you using `make test`?

Re replica sets: https://docs.mongodb.org/v3.0/tutorial/deploy-replica-set/